refactor: [LAUNCHER-BUG]: Flatpak login still doesnt work (Linux not found error?) #2342 https://github.com/NoRiskClient/issues/issues/2342
